requirements.txt
openai
langchain
langchain_openai
langchain_experimental
langchainhub
#pinecone-client
python-dotenv
qrcode
#tiktoken
#docx2txt
#pypdf
#requests
#numpy
pandas
tabulate
#wikipedia
#streamlit
#streamlit_chat
#pexpect